CDI controller code dedupe
