+++ pkg/controller/clone-controller_test.go
@@ -7,85 +7,44 @@ import (
 
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/apimachinery/pkg/util/diff"
-	kubeinformers "k8s.io/client-go/informers"
-	k8sfake "k8s.io/client-go/kubernetes/fake"
-	core "k8s.io/client-go/testing"
 	"k8s.io/client-go/tools/cache"
 	. "kubevirt.io/containerized-data-importer/pkg/common"
 )
 
-type cloneFixture struct {
-	t *testing.T
-
-	kubeclient *k8sfake.Clientset
-
-	// Objects to put in the store.
-	pvcLister []*corev1.PersistentVolumeClaim
-	podLister []*corev1.Pod
-
-	// Actions expected to happen on the client.
-	kubeactions []core.Action
-	actions     []core.Action
-
-	// Objects from here preloaded into NewSimpleFake.
-	kubeobjects []runtime.Object
+type CloneFixture struct {
+	ControllerFixture
 }
 
-func newCloneFixture(t *testing.T) *cloneFixture {
-	f := &cloneFixture{}
-	f.t = t
-	f.kubeobjects = []runtime.Object{}
+func newCloneFixture(t *testing.T) *CloneFixture {
+	f := &CloneFixture{
+		ControllerFixture: *newControllerFixture(t),
+	}
 	return f
 }
 
-func (f *cloneFixture) newController() *CloneController {
-
-	f.kubeclient = k8sfake.NewSimpleClientset(f.kubeobjects...)
-
-	podFactory := kubeinformers.NewSharedInformerFactory(f.kubeclient, noResyncPeriodFunc())
-	pvcFactory := kubeinformers.NewSharedInformerFactory(f.kubeclient, noResyncPeriodFunc())
-
-	podInformer := podFactory.Core().V1().Pods()
-	pvcInformer := pvcFactory.Core().V1().PersistentVolumeClaims()
-
-	c := NewCloneController(f.kubeclient,
-		pvcInformer,
-		podInformer,
-		"test/mycloneimage",
-		"Always",
-		"5")
-
-	for _, pod := range f.podLister {
-		c.podInformer.GetIndexer().Add(pod)
+func (f *CloneFixture) newCloneController() *CloneController {
+	return &CloneController{
+		Controller: *f.newController("test/mycloneimage", "Always", "5"),
 	}
-
-	for _, pvc := range f.pvcLister {
-		c.pvcInformer.GetIndexer().Add(pvc)
-	}
-
-	return c
 }
 
-func (f *cloneFixture) run(pvcName string) {
+func (f *CloneFixture) run(pvcName string) {
 	f.runController(pvcName, true, false, false)
 }
 
-func (f *cloneFixture) runWithExpectation(pvcName string) {
+func (f *CloneFixture) runWithExpectation(pvcName string) {
 	f.runController(pvcName, true, false, true)
 }
 
-func (f *cloneFixture) runExpectError(pvcName string) {
+func (f *CloneFixture) runExpectError(pvcName string) {
 	f.runController(pvcName, true, true, false)
 }
 
-func (f *cloneFixture) runController(pvcName string,
+func (f *CloneFixture) runController(pvcName string,
 	startInformers bool,
 	expectError bool,
 	withCreateExpectation bool) {
-	c := f.newController()
+	c := f.newCloneController()
 	if startInformers {
 		stopCh := make(chan struct{})
 		defer close(stopCh)
@@ -105,7 +64,7 @@ func (f *cloneFixture) runController(pvcName string,
 		f.t.Error("expected error syncing pvc, got nil")
 	}
 
-	k8sActions := filterCloneActions(f.kubeclient.Actions())
+	k8sActions := filterActions(f.kubeclient.Actions())
 	for i, action := range k8sActions {
 		if len(f.kubeactions) < i+1 {
 			f.t.Errorf("%d unexpected actions: %+v", len(k8sActions)-len(f.kubeactions), k8sActions[i:])
@@ -113,83 +72,12 @@ func (f *cloneFixture) runController(pvcName string,
 		}
 
 		expectedAction := f.kubeactions[i]
-		checkCloneAction(expectedAction, action, f.t)
+		checkAction(expectedAction, action, f.t)
 	}
 
 	if len(f.kubeactions) > len(k8sActions) {
 		f.t.Errorf("%d additional expected actions:%+v", len(f.kubeactions)-len(k8sActions), f.kubeactions[len(k8sActions):])
 	}
-}
-
-// checkCloneAction verifies that expected and actual actions are equal and both have
-// same attached resources
-func checkCloneAction(expected, actual core.Action, t *testing.T) {
-	if !(expected.Matches(actual.GetVerb(), actual.GetResource().Resource) && actual.GetSubresource() == expected.GetSubresource()) {
-		t.Errorf("Expected\n\t%#v\ngot\n\t%#v", expected, actual)
-		return
-	}
-
-	if reflect.TypeOf(actual) != reflect.TypeOf(expected) {
-		t.Errorf("Action has wrong type. Expected: %t. Got: %t", expected, actual)
-		return
-	}
-
-	switch a := actual.(type) {
-	case core.CreateAction:
-		e, _ := expected.(core.CreateAction)
-		expObject := e.GetObject()
-		object := a.GetObject()
-
-		if !reflect.DeepEqual(expObject, object) {
-			t.Errorf("Action %s %s has wrong object\nDiff:\n %s",
-				a.GetVerb(), a.GetResource().Resource, diff.ObjectGoPrintDiff(expObject, object))
-		}
-	case core.UpdateAction:
-		e, _ := expected.(core.UpdateAction)
-		expObject := e.GetObject()
-		object := a.GetObject()
-
-		if !reflect.DeepEqual(expObject, object) {
-			t.Errorf("Action %s %s has wrong object\nDiff:\n %s",
-				a.GetVerb(), a.GetResource().Resource, diff.ObjectGoPrintDiff(expObject, object))
-		}
-	case core.PatchAction:
-		e, _ := expected.(core.PatchAction)
-		expPatch := e.GetPatch()
-		patch := a.GetPatch()
-
-		if !reflect.DeepEqual(expPatch, expPatch) {
-			t.Errorf("Action %s %s has wrong patch\nDiff:\n %s",
-				a.GetVerb(), a.GetResource().Resource, diff.ObjectGoPrintDiff(expPatch, patch))
-		}
-	}
-}
-
-// filterCloneActions filters list and watch actions for testing resources.
-// Since list and watch don't change resource state we can filter it to lower
-// nose level in our tests.
-func filterCloneActions(actions []core.Action) []core.Action {
-	ret := []core.Action{}
-	for _, action := range actions {
-		if len(action.GetNamespace()) == 0 &&
-			(action.Matches("list", "persistentvolumeclaims") ||
-				action.Matches("watch", "persistentvolumeclaims") ||
-				action.Matches("list", "pods") ||
-				action.Matches("watch", "pods")) {
-			continue
-		}
-		ret = append(ret, action)
-	}
-
-	return ret
-}
-
-func (f *cloneFixture) expectCreatePodAction(d *corev1.Pod) {
-	f.kubeactions = append(f.kubeactions, core.NewCreateAction(schema.GroupVersionResource{Resource: "pods", Version: "v1"}, d.Namespace, d))
-}
-
-func (f *cloneFixture) expectUpdatePvcAction(d *corev1.PersistentVolumeClaim) {
-	f.kubeactions = append(f.kubeactions, core.NewUpdateAction(schema.GroupVersionResource{Resource: "persistentvolumeclaims", Version: "v1"}, d.Namespace, d))
 }
 
 // Verifies basic pods creation when new PVC is discovered
@@ -361,7 +249,7 @@ func TestCloneFindPodsInCacheUpdating(t *testing.T) {
 	f.kubeobjects = append(f.kubeobjects, tests[1].sourcePod)
 	f.kubeobjects = append(f.kubeobjects, tests[1].targetPod)
 
-	controller := f.newController()
+	controller := f.newCloneController()
 
 	stopCh := make(chan struct{})
 	defer close(stopCh)
