+++ pkg/controller/datavolume-controller_test.go
@@ -146,7 +146,7 @@ func (f *fixture) runController(dataVolumeName string, startInformers bool, expe
 		}
 
 		expectedAction := f.actions[i]
-		checkAction(expectedAction, action, f.t)
+		checkDVAction(expectedAction, action, f.t)
 	}
 
 	if len(f.actions) > len(actions) {
@@ -161,7 +161,7 @@ func (f *fixture) runController(dataVolumeName string, startInformers bool, expe
 		}
 
 		expectedAction := f.kubeactions[i]
-		checkAction(expectedAction, action, f.t)
+		checkDVAction(expectedAction, action, f.t)
 	}
 
 	if len(f.kubeactions) > len(k8sActions) {
@@ -169,9 +169,9 @@ func (f *fixture) runController(dataVolumeName string, startInformers bool, expe
 	}
 }
 
-// checkAction verifies that expected and actual actions are equal and both have
+// checkDVAction verifies that expected and actual actions are equal and both have
 // same attached resources
-func checkAction(expected, actual core.Action, t *testing.T) {
+func checkDVAction(expected, actual core.Action, t *testing.T) {
 	if !(expected.Matches(actual.GetVerb(), actual.GetResource().Resource) && actual.GetSubresource() == expected.GetSubresource()) {
 		t.Errorf("Expected\n\t%#v\ngot\n\t%#v", expected, actual)
 		return
